Only install packages if we've got a pyproject.toml

+++ students/start_notebook_env.sh
@@ -6,10 +6,13 @@ if ! grep -qi 'amzn' /etc/os-release; then
 else
   uv --version || (curl -LsSf https://astral.sh/uv/install.sh | sh)
   echo "AWS" > "/tmp/runtime"
-  # Generate requirements.txt from pyproject.toml using uv
-  /home/ec2-user/.local/bin/uv pip compile pyproject.toml -o requirements.txt --quiet
-  # Then install normally
-  pip install -q -r requirements.txt
+  
+  if [ -f pyproject.toml ]; then
+    echo "pyproject.toml found. Generating requirements.txt and installing dependencies. This may take a few minutes..."
+    /home/ec2-user/.local/bin/uv pip compile pyproject.toml -o requirements.txt --quiet
+    pip install -q -r requirements.txt
+  fi
+  
   if ! sudo docker ps --format '{{.Names}}' | grep -q '^ollama$'; then
     sudo docker run -d --name ollama -p 11434:11434 -v "$HOME/.ollama":/root/.ollama ollama/ollama:latest
   fi
